+++ NodeLogin/package.json
@@ -1,0 +1,40 @@
+{
+    "name": "authenticator",
+    "version": "1.0.0",
+    "description": "",
+    "main": "server.js",
+    "scripts": {
+        "test": "echo \"Error: no test specified\" && exit 1"
+    },
+    "author": "",
+    "license": "ISC",
+    "dependencies": {
+        "babel-preset-es2015": "^6.24.1",
+        "babel-preset-stage-0": "^6.24.1",
+        "bcrypt": "^3.0.2",
+        "body-parser": "^1.18.3",
+        "cors": "^2.8.4",
+        "dotenv": "^6.1.0",
+        "express": "^4.17.1",
+        "jsonwebtoken": "^8.3.0",
+        "morgan": "^1.9.0",
+        "mysql2": "^1.5.3",
+        "nodemailer": "^4.6.8",
+        "nodemon": "^1.18.3",
+        "passport": "^0.4.0",
+        "passport-jwt": "^4.0.0",
+        "passport-local": "^1.0.0",
+        "sequelize": "^5.3.0",
+        "sequelize-cli": "^5.5.0",
+        "swagger-jsdoc": "^3.2.9",
+        "swagger-ui-express": "^4.0.2"
+    },
+    "devDependencies": {
+        "@babel/cli": "^7.5.5",
+        "@babel/core": "^7.5.5",
+        "@babel/node": "^7.5.5",
+        "@babel/preset-env": "^7.5.5",
+        "@babel/register": "^7.5.5",
+        "babel-cli": "^6.26.0"
+    }
+}
