--- config/jwtConfig.js
@@ -1,3 +1,0 @@
-module.exports = {
-    secret: 'jwt-secret',
-}
